+++ app/models/daos/DaoContext.scala
@@ -1,0 +1,8 @@
+package models.daos
+
+import javax.inject._
+
+@Singleton
+class DaoContext @Inject() (
+  val userDao: UserDao,
+  val authTokenDao: AuthTokenDao)
