+++ app/models/daos/UserDao.scala
@@ -4,15 +4,13 @@ import java.util.UUID
 
 import com.mohiva.play.silhouette.api.LoginInfo
 import models.User
-import models.daos.UserDAOImpl._
 
-import scala.collection.mutable
 import scala.concurrent.Future
 
 /**
  * Give access to the user object.
  */
-class UserDAOImpl extends UserDAO {
+trait UserDao {
 
   /**
    * Finds a user by its login info.
@@ -20,9 +18,7 @@ class UserDAOImpl extends UserDAO {
    * @param loginInfo The login info of the user to find.
    * @return The found user or None if no user for the given login info could be found.
    */
-  def find(loginInfo: LoginInfo) = Future.successful(
-    users.find { case (_, user) => user.loginInfo == loginInfo }.map(_._2)
-  )
+  def find(loginInfo: LoginInfo): Future[Option[User]]
 
   /**
    * Finds a user by its user ID.
@@ -30,7 +26,7 @@ class UserDAOImpl extends UserDAO {
    * @param userID The ID of the user to find.
    * @return The found user or None if no user for the given ID could be found.
    */
-  def find(userID: UUID) = Future.successful(users.get(userID))
+  def find(userID: UUID): Future[Option[User]]
 
   /**
    * Saves a user.
@@ -38,19 +34,5 @@ class UserDAOImpl extends UserDAO {
    * @param user The user to save.
    * @return The saved user.
    */
-  def save(user: User) = {
-    users += (user.userID -> user)
-    Future.successful(user)
-  }
-}
-
-/**
- * The companion object.
- */
-object UserDAOImpl {
-
-  /**
-   * The list of users.
-   */
-  val users: mutable.HashMap[UUID, User] = mutable.HashMap()
+  def save(user: User): Future[User]
 }
