+++ app/models/daos/UserDaoImpl.scala
@@ -4,13 +4,15 @@ import java.util.UUID
 
 import com.mohiva.play.silhouette.api.LoginInfo
 import models.User
+import models.daos.UserDaoImpl._
 
+import scala.collection.mutable
 import scala.concurrent.Future
 
 /**
  * Give access to the user object.
  */
-trait UserDAO {
+class UserDaoImpl extends UserDao {
 
   /**
    * Finds a user by its login info.
@@ -18,7 +20,9 @@ trait UserDAO {
    * @param loginInfo The login info of the user to find.
    * @return The found user or None if no user for the given login info could be found.
    */
-  def find(loginInfo: LoginInfo): Future[Option[User]]
+  def find(loginInfo: LoginInfo) = Future.successful(
+    users.find { case (_, user) => user.loginInfo == loginInfo }.map(_._2)
+  )
 
   /**
    * Finds a user by its user ID.
@@ -26,7 +30,7 @@ trait UserDAO {
    * @param userID The ID of the user to find.
    * @return The found user or None if no user for the given ID could be found.
    */
-  def find(userID: UUID): Future[Option[User]]
+  def find(userID: UUID) = Future.successful(users.get(userID))
 
   /**
    * Saves a user.
@@ -34,5 +38,19 @@ trait UserDAO {
    * @param user The user to save.
    * @return The saved user.
    */
-  def save(user: User): Future[User]
+  def save(user: User) = {
+    users += (user.userID -> user)
+    Future.successful(user)
+  }
+}
+
+/**
+ * The companion object.
+ */
+object UserDaoImpl {
+
+  /**
+   * The list of users.
+   */
+  val users: mutable.HashMap[UUID, User] = mutable.HashMap()
 }
