+++ app/modules/BaseModule.scala
@@ -1,7 +1,7 @@
 package modules
 
 import com.google.inject.AbstractModule
-import models.daos.{ AuthTokenDAO, AuthTokenDAOImpl }
+import models.daos.{ AuthTokenDao, AuthTokenDAOImpl }
 import models.services.{ AuthTokenService, AuthTokenServiceImpl }
 import net.codingwell.scalaguice.ScalaModule
 
@@ -13,7 +13,7 @@ class BaseModule extends AbstractModule with ScalaModule {
    * Configures the module.
    */
   override def configure(): Unit = {
-    bind[AuthTokenDAO].to[AuthTokenDAOImpl]
+    bind[AuthTokenDao].to[AuthTokenDAOImpl]
     bind[AuthTokenService].to[AuthTokenServiceImpl]
   }
 }
